+++ src/main/java/me/jinwenjie/dao/LabelDao.java
@@ -3,13 +3,13 @@ package me.jinwenjie.dao;
 import me.jinwenjie.model.Label;
 
 public interface LabelDao {
-    int deleteByPrimaryKey(Long labelId);
+    int deleteByPrimaryKey(Integer labelId);
 
     int insert(Label record);
 
     int insertSelective(Label record);
 
-    Label selectByPrimaryKey(Long labelId);
+    Label selectByPrimaryKey(Integer labelId);
 
     int updateByPrimaryKeySelective(Label record);
